mqtt_influx_telegraf_emulator/requirements.txt
paho-mqtt
influxdb-client
schedule
pyyaml